pyproject.toml
[project]
name = "docurag-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "inngest>=0.5.9",
    "llama-index-core>=0.14.3",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
    "pymongo>=4.6.0",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
    "nest-asyncio>=1.6.0",
    "python-multipart>=0.0.20",
    "boto3>=1.42.5",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
